Add files via upload
The idea behind this change is to start the newly generated story with a name of a prominent marvel character. Hopefully, this will make the new stories more intelligible and closer to actual storylines. 

To do this, I scraped the first 5 characters in the character list of the IMDB webpage for each Marvel movie. These are cleaned up and saved as a file, then imported into the run.py file. They are then used, in conjunction with the markov chain functions to generate a new storyline. The Markov Chain functions work very similar to the original program, with some changes to the definition and origin of the first word (see comments within document).